Move fx example below fx package
