+++ src/main/java/com/koreaap/itdev2/fx/Pair.java
@@ -1,4 +1,4 @@
-package com.koreaap.itdev2;
+package com.koreaap.itdev2.fx;
 
 public class Pair {
   private String from;
